LayoutTests/media/video-loopstart.html

<video playcount=2 loopstart=1s controls></video>
<p>Test that playing video rewinds to time specified by the loopstart attribute when it loops.</p>
<script src=video-test.js></script>
<script>
    testExpected("video.loopStart", 1.0);

    waitForEvent('load', function () {
        waitForEvent('timeupdate', function () {
            failTestIn(1000);
            if (video.currentLoop == 1) {
                testExpected("video.currentLoop", 1);

                testExpected("video.currentTime.toFixed(1)", 1.0);

                endTest();
            }
        });
        video.currentTime = video.duration - 0.2;
    });

    video.src = "content/test.mp4";

    run("video.play()");

</script>
